refactor: move is_remote_url() to test_runner.rs
